create a new component UploadFeaturedImage and move AddImage to lib
